CMakeLists.txt: fix hang when >1 Camomile LV2 plugin is loaded (issue #286)
This issue was ultimately caused by an object buried in the JUCE code having
an inappropriate scope - GCC made the object a "unique global symbol", which
means there is only one instance across the current process.
When multiple plugins were loaded, they all shared this instance, but did
not share other relevant objects (from the JUCE code), causing havoc.
The fix is to ban GCC from marking symbols as "unique global".

Also set the C++ standard to be C++20 for the Camomile_LV2 meta-plugin, in
line with the other meta-plugins.